Finish test setup
