+++ op-e2e/e2eutils/wait/blocks.go
@@ -7,6 +7,7 @@ import (
 	"time"
 
 	"github.com/ethereum/go-ethereum/core/types"
+	"github.com/ethereum/go-ethereum/rpc"
 
 	"github.com/ethereum-optimism/optimism/op-service/eth"
 	"github.com/ethereum-optimism/optimism/op-service/sources"
@@ -36,21 +37,6 @@ func ForBlock(ctx context.Context, client BlockCaller, n uint64) error {
 			return nil
 		}
 	}
-	// for {
-	// 	if ctx.Done() != nil {
-	// 		return ctx.Err()
-	// 	}
-	// 	height, err := client.BlockNumber(ctx)
-	// 	if err != nil {
-	// 		return err
-	// 	}
-	// 	if height < n {
-	// 		time.Sleep(500 * time.Millisecond)
-	// 		continue
-	// 	}
-	// 	break
-	// }
-	// return nil
 }
 
 func ForBlockWithTimestamp(ctx context.Context, client BlockCaller, target uint64) error {
@@ -84,4 +70,31 @@ func ForProcessingFullBatch(ctx context.Context, rollupCl *sources.RollupClient)
 		return syncStatus.PendingSafeL2 == syncStatus.SafeL2
 	})
 	return err
+}
+
+func ForNextSafeBlock(ctx context.Context, client BlockCaller) error {
+	safeBlockNumber := big.NewInt(rpc.SafeBlockNumber.Int64())
+	current, err := client.BlockByNumber(ctx, safeBlockNumber)
+	if err != nil {
+		return err
+	}
+
+	// Long timeout so we don't have to care what the block time is. If the test passes this will complete early anyway.
+	ctx, cancel := context.WithTimeout(ctx, 60*time.Second)
+	defer cancel()
+	for {
+		select {
+		case <-ctx.Done():
+			return ctx.Err()
+		default:
+			next, err := client.BlockByNumber(ctx, safeBlockNumber)
+			if err != nil {
+				return err
+			}
+			if next.NumberU64() > current.NumberU64() {
+				return nil
+			}
+			time.Sleep(500 * time.Millisecond)
+		}
+	}
 }
